LayoutTests/http/tests/security/mixedContent/redirect-http-to-https-iframe-in-main-frame-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-redirect-http-to-https-frame.html requested insecure content from http://127.0.0.1:8080/resources/redirect.py?url=https://127.0.0.1:8443/security/mixedContent/resources/boring.html. This content was blocked and must be served over HTTPS.

PASS window.data is 'blocked'
This test opens a window that loads an insecure iframe (via a tricky redirect). We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying content that can be controlled by an active network attacker.
